+++ src/components/oauth/NaverLogin.tsx
@@ -1,32 +1,32 @@
 import { useState, useEffect } from 'react';
-import { resetAuthManager } from '../auth/authManager';
-import { generateRandomString, generateCodeVerifier, generateCodeChallenge } from '../utils/pkceUtils';
+import { resetAuthManager } from '../../auth/authManager';
+import { generateRandomString, generateCodeVerifier, generateCodeChallenge } from '../../utils/pkceUtils';
 
-const KakaoLogin = () => {
+const NaverLogin = () => {
   const [isLoading, setIsLoading] = useState(false);
   const [message, setMessage] = useState('');
 
   // AuthManager 초기화
   useEffect(() => {
-    resetAuthManager('kakao');
+    resetAuthManager('naver');
     setMessage('');
   }, []);
 
-  const handleKakaoLogin = async () => {
+  const handleNaverLogin = async () => {
     try {
       setIsLoading(true);
-      setMessage('🔄 Kakao OAuth 페이지로 이동합니다...');
+      setMessage('🔄 Naver OAuth 페이지로 이동합니다...');
       
-      // 환경변수에서 Kakao OAuth 설정 가져오기
-      const clientId = import.meta.env.VITE_KAKAO_CLIENT_ID; 
-      const redirectUri = `${window.location.origin}/auth/kakao/callback`;
+      // 환경변수에서 Naver OAuth 설정 가져오기
+      const clientId = import.meta.env.VITE_NAVER_CLIENT_ID; 
+      const redirectUri = `${window.location.origin}/auth/naver/callback`;
       
       // 환경변수 redirect URI가 있으면 사용, 없으면 기본값 사용
-      const finalRedirectUri = import.meta.env.VITE_KAKAO_REDIRECT_URI || redirectUri;
+      const finalRedirectUri = import.meta.env.VITE_NAVER_REDIRECT_URI || redirectUri;
       
       // 환경변수 검증
       if (!clientId) {
-        setMessage('❌ Kakao Client ID가 설정되지 않았습니다. .env 파일에 VITE_KAKAO_CLIENT_ID를 설정해주세요.');
+        setMessage('❌ Naver Client ID가 설정되지 않았습니다. .env 파일에 VITE_NAVER_CLIENT_ID를 설정해주세요.');
         setIsLoading(false);
         return;
       }
@@ -37,30 +37,30 @@ const KakaoLogin = () => {
       const state = generateRandomString(32);
       
       // PKCE 파라미터를 localStorage에 저장
-      localStorage.setItem('kakao_oauth_code_verifier', codeVerifier);
-      localStorage.setItem('kakao_oauth_state', state);
+      localStorage.setItem('naver_oauth_code_verifier', codeVerifier);
+      localStorage.setItem('naver_oauth_state', state);
       
-      // Kakao OAuth URL 생성 (PKCE 포함)
-      const kakaoAuthUrl = `https://kauth.kakao.com/oauth/authorize?` +
+      // Naver OAuth URL 생성 (PKCE 포함)
+      const naverAuthUrl = `https://nid.naver.com/oauth2.0/authorize?` +
         `client_id=${encodeURIComponent(clientId)}&` +
         `redirect_uri=${encodeURIComponent(finalRedirectUri)}&` +
         `response_type=code&` +
-        `scope=${encodeURIComponent('profile_nickname account_email')}&` +
+        `scope=${encodeURIComponent('name email')}&` +
         `code_challenge=${encodeURIComponent(codeChallenge)}&` +
         `code_challenge_method=S256&` +
         `state=${encodeURIComponent(state)}`;
       
       // 현재 상태를 localStorage에 저장
       localStorage.setItem('oauth_in_progress', 'true');
-      localStorage.setItem('oauth_provider', 'kakao');
+      localStorage.setItem('oauth_provider', 'naver');
       
-      // 현재 창에서 Kakao OAuth 페이지로 이동
-      window.location.href = kakaoAuthUrl;
+      // 현재 창에서 Naver OAuth 페이지로 이동
+      window.location.href = naverAuthUrl;
       
     } catch (error) {
       const errorMessage = error instanceof Error ? error.message : '알 수 없는 오류';
       setMessage(`❌ ${errorMessage}`);
-      console.error('❌ Kakao 로그인 중 오류:', error);
+      console.error('❌ Naver 로그인 중 오류:', error);
       setIsLoading(false);
     }
   };
@@ -71,8 +71,8 @@ const KakaoLogin = () => {
 
       {/* 헤더 */}
       <div className="px-4 py-16 pb-4 text-center">
-        <h2 className="text-2xl font-bold text-gray-900 mb-2">Kakao로 계속하기</h2>
-        <p className="text-sm text-gray-600">Kakao 계정으로 빠르게 로그인하세요</p>
+        <h2 className="text-2xl font-bold text-gray-900 mb-2">Naver로 계속하기</h2>
+        <p className="text-sm text-gray-600">Naver 계정으로 빠르게 로그인하세요</p>
       </div>
 
       <div className="flex-1 px-8 pb-4 flex flex-col w-full">
@@ -90,11 +90,11 @@ const KakaoLogin = () => {
             )}
 
             <button
-              onClick={handleKakaoLogin}
+              onClick={handleNaverLogin}
               disabled={isLoading}
-              className="w-full p-4 bg-yellow-400 text-black rounded-xl text-base font-semibold hover:bg-yellow-500 hover:-translate-y-0.5 transition-all duration-200 disabled:opacity-50 disabled:cursor-not-allowed disabled:transform-none"
+              className="w-full p-4 bg-green-600 text-white rounded-xl text-base font-semibold hover:bg-green-700 hover:-translate-y-0.5 transition-all duration-200 disabled:opacity-50 disabled:cursor-not-allowed disabled:transform-none"
             >
-              {isLoading ? 'Kakao 인증 중...' : 'Kakao로 계속하기'}
+              {isLoading ? 'Naver 인증 중...' : 'Naver로 계속하기'}
             </button>
           </div>
         </div>
@@ -103,4 +103,4 @@ const KakaoLogin = () => {
   );
 };
 
-export default KakaoLogin;
+export default NaverLogin;
